some cleanup and migration for the Registration entity
