+++ src/modules/characters/Modules.Characters.Endpoints/Entities/Characters/Create/CreateCharacterRequest.cs
@@ -1,0 +1,3 @@
+﻿namespace Starlights.Modules.Characters.Endpoints.Entities.Characters.Create;
+
+public record CreateCharacterRequest(Guid CharacterCreationOptionId, string Name, string? PortraitUrl = null);
