+++ src/modules/characters/Modules.Characters.Endpoints/Entities/Characters/Create/CreateCharacterResponse.cs
@@ -1,0 +1,3 @@
+﻿namespace Starlights.Modules.Characters.Endpoints.Entities.Characters.Create;
+
+public record CreateCharacterResponse(Guid Id);
